requirements.txt
python-dotenv==1.0.0
slack-bolt==1.18.0
slack-sdk==3.21.3
